Fixed bug with m2m without intermediary model backends not being executed

+++ orchestra/apps/resources/forms.py
@@ -7,7 +7,7 @@ from orchestra.forms.widgets import ShowTextWidget, ReadOnlyWidget
 class ResourceForm(forms.ModelForm):
     verbose_name = forms.CharField(label=_("Name"), required=False,
             widget=ShowTextWidget(bold=True))
-    allocated = forms.IntegerField(label=_("Allocated"))
+    allocated = forms.DecimalField(label=_("Allocated"))
     unit = forms.CharField(label=_("Unit"), widget=ShowTextWidget(), required=False)
     
     class Meta:
